% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barinfo.R
\name{infobar}
\alias{infobar}
\title{Title}
\usage{
infobar(
  n,
  cases_right,
  cases_left,
  outcome_texts,
  headline_1_text_left,
  headline_2_text_left,
  headline_2_text_right,
  center_distance = 0.01,
  fontsize = 1,
  description_size = 0.6,
  bar_size = 0.4,
  n_size = 3.5 * fontsize,
  n_fontface = "bold",
  col_left = rgb(0.3, 0.3, 0.3),
  col_right = "#D17A00",
  headline1_fontface = "plain",
  headline2_fontface = "bold",
  n_height = 0.7,
  headline1_height = 0.37,
  headline2_height = 0.2,
  headline_1_text_right = headline_1_text_left,
  headline_size = 1.5 * fontsize,
  height_first = 0.9,
  distance_box = 0.4,
  box_height = 0.4,
  case_distance = 0.01,
  case_size = 1.8 * fontsize,
  outcome_col = rgb(0.3, 0.3, 0.3),
  outcome_size = 1.5 * fontsize,
  outcome_fontface = "bold",
  outcome_cases_fontface = "bold"
)
}
\arguments{
\item{outcome_cases_fontface}{}
}
\value{

}
\description{
Title
}
